Fixed bug with bad attribute keys and access to member methods.

+++ test/fixtures/pointers/Size.h
@@ -11,7 +11,7 @@ struct Size {
   Size(double tx, double ty)
     : x(tx)
     , y(ty) {}
-  float surface() {
+  double surface() {
     return x * y;
   }
 };
